Codigo/Backend/requirements.txt
asgiref==3.7.2
Brotli==1.1.0
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
mysqlclient==2.2.1
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
whitenoise==6.6.0
